price the fractional fill sizes the wire actually sends

+++ bot/lag/fee_floor.py
@@ -21,9 +21,10 @@ class FeeSource:
 
 
 # Deliberately duplicates fees.taker_fee rather than delegating to it: this is the oracle that
-# module is checked against, and delegating would make fee_module_corrected a tautology.
-def published_taker_fee(contracts: int, price: Decimal) -> Decimal:
-    raw = PUBLISHED_TAKER_RATE * Decimal(contracts) * price * (Decimal("1") - price)
+# module is checked against, and delegating would make fee_module_corrected a tautology. It takes
+# the size as a Decimal because the wire prints fractional counts that fees.taker_fee cannot say.
+def published_taker_fee(contracts: Decimal, price: Decimal) -> Decimal:
+    raw = PUBLISHED_TAKER_RATE * contracts * price * (Decimal("1") - price)
     return raw.quantize(CENT, rounding=ROUND_CEILING)
